Update to use separate polyfills file
