+++ src/polyfills.deno.ts
@@ -1,0 +1,13 @@
+export var promiseImpl = Promise
+
+export function getAsyncFunctionConstructor (): Function {
+  return async function () {}.constructor
+}
+
+export function trimLeft (str: string): string {
+  return str.trimLeft()
+}
+
+export function trimRight (str: string): string {
+  return str.trimRight()
+}
